Django/장고 1018/1018/practice/requirements.txt
asgiref==3.5.2
Django==3.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==15.1.1
psycopg2==2.9.4
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
toposort==1.7
